% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfdct-package.R
\name{antarctica}
\alias{antarctica}
\title{Antarctica, and not Antarctica.}
\description{
See data-raw for original source.
}
\examples{
library(sf)
plot(antarctica, col = rainbow(nrow(antarctica), alpha = 0.4))
}
