% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfdct-package.R
\docType{package}
\name{sfdct-package}
\alias{sfdct}
\alias{sfdct-package}
\title{sfdct: Constrained Triangulation for Simple Features}
\description{
Build a constrained high quality Delaunay triangulation from simple features objects, applying constraints based on input line segments, and triangle properties including maximum area, minimum internal angle. The triangulation code in 'RTriangle' uses the method of Cheng, Dey and Shewchuk (2012, ISBN:9781584887300). For a low-dependency alternative with low-quality path-based constrained triangulation see \url{https://CRAN.R-project.org/package=decido} and for high-quality configurable triangulation see \url{https://github.com/hypertidy/anglr}. Also consider comparison with the 'GEOS' lib which since version 3.10.0 includes a low quality polygon triangulation method that starts with ear clipping and refines to Delaunay.

Triangulate simple features including the input edges as constraints, rather than
being bounded to the convex hull.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/sfdct}
  \item Report bugs at \url{https://github.com/hypertidy/sfdct/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
